deps: Upgrade ncc to 0.34.0 to resolve OpenSSL issue.
At least some Node v16 installations are unable to build this package
due to "unsupported" "digital envelope routines". Based on a
StackOverflow answer [^1], this appears to be OpenSSL-related, and also
somewhat version and/or distribution specific (for example, Void Linux
does not repro the issue). Upgrading @vercel/ncc to 0.34 resolves the
issue in Anders's testing, so we'll take that upgrade here, and
regenerate the dist file using this new version.

[^1]: https://stackoverflow.com/a/73027407